Fix for wrap_layout bug
When changing over to Middleman 3.3.3 I noticed that there was an error:

NoMethodError: undefined method 'safe_concat' for "":String

This was talked about in the following GitHub issue:

https://github.com/middleman/middleman/issues/1269

I updated all the slim templates to use the '=' operator to ensure their
output. The `concept.erb` template would not work with the `=` operator
and without it displayed no text, so I replaced it with a slim version of
the template.