Fix for tab helpers to properly surround with section tag
This helper previously surrounded the title and content with
a section tag. With the upgrade that seemed to no longer work
because of blocks being used. I opted instead to use the second
parameter as content and not options and forgo the block. This
works to properly encapsulate the content.