Adds the language as a class to the html output
This adds the ability to build custom styles per type of code output as
the highlight language is added as a class to the surrounding div.

Example:

```
<div class="highlight shell">
  <pre>knife cookbook create iis-tutorial-1</pre>
</div>
```

Upgraded to the latest middleman-syntax which changed over to rouge.
Rouge changed the wrapping format from pygments so the easist way was
to disable the current wrapping and override syntax renderer. This required
a monkeypatch. Ugh!