set disqus_url
In working with Disqus’ support team, they suggested that we set the
disqus_url variable so that:

1. The proper canonical is used the first time a page is visited and
the thread is created.
2. Ensure that if the non-canonical URL is used (either by accident or
on purpose) the correct thread is associated with the page.

I ensured that the URL ends with a / because we redirect to that if
it’s not provided.

This may or may not fix the issue we’ve been having with threads not
appearing, but this is a best practice that they recommend doing anyway.