Adding the Joy of Automating Ep. 1 and Ep. 2
Up to this point I have been using a custom repository to store the Joy of Automating activites and resources. I want to make sure that this content is captured in the learn chef Skills Library. I also want to be able to determine if people are actually visiting the activity page and I cannot do that with GitHub location. This has the benefit of comments and other things as well.

Overall it is the right move.