Ruby behind Chef - Removes the TODO
The string example is simple but I don't want to complicate it with
way more to get the point across.
The Array section is long enough and while I would like to talk about
adding / subtracting / compacting / and `uniq`ing them I am opting not
to make that section any longer.

Signed-off-by: Franklin Webber <franklin@chef.io>