Restore Ruby script line numbers
Line numbers were missing because this is considered a legacy feature of the Rouge gem, which can fortunately be easily re-enabled. There was also an issue where <pre> tags were being confused with paragraph tags in the Angularize helpers, which was breaking the formatting of line numbers once they were re-enabled.