Adds some of the PowerShell page
I don't know much about the venacular about PowerShell so I
had to spend a fair amount of time researching more about it.
I know that we want to create a resource to help people understand
it but I don't think this is in the state that would be useful
for someone yet.

Right now it's really very basic and though it's about the length
of the Linux section I feel like it doesn't do enough to help
someone really know about PowerShell.

I think if I were to continue on this section I would suggest
cheating a bit and using a few examples that the learner would
see in some of the modules where they use PowerShell.

Signed-off-by: Franklin Webber <franklin@chef.io>